added attributes to submitted piece for faster loading
